Only send params that are used to the database
This is prep work to allow bringing larger datasets into scope in a
script, without disrupting the query performance by sending it off
to the server.

This is similar to how pgbench works; pgbench does the substitution
on the client side, so gets this behavior "for free", so to speak.